requirements.txt
streamlit==1.55.0
langchain==1.2.13
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-google-genai==4.2.1
faiss-cpu==1.13.2
pypdf==6.9.2
python-docx==1.2.0
python-dotenv==1.2.2
